+++ src/app/components/content-atm-inventories-workflows.js
@@ -9,20 +9,30 @@
 
 import Component from '@ember/component';
 import { inject as service } from '@ember/service';
-import { conditional, array, promise } from 'ember-awesome-macros';
+import { promise } from 'ember-awesome-macros';
 import EmberObject, { computed, observer, get } from '@ember/object';
 import { reads } from '@ember/object/computed';
 import GlobalActions from 'onedata-gui-common/mixins/components/global-actions';
 import ActionsFactory from 'onedata-gui-common/utils/workflow-visualiser/actions-factory';
 import { Promise } from 'rsvp';
 import { scheduleOnce } from '@ember/runloop';
+import preventPageUnload from 'onedata-gui-common/utils/prevent-page-unload';
+import I18n from 'onedata-gui-common/mixins/components/i18n';
+import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 
-export default Component.extend(GlobalActions, {
+export default Component.extend(GlobalActions, I18n, {
   classNames: ['content-atm-inventories-workflows'],
 
   navigationState: service(),
   recordManager: service(),
   workflowManager: service(),
+  router: service(),
+  modalManager: service(),
+
+  /**
+   * @override
+   */
+  i18nPrefix: 'components.contentAtmInventoriesWorkflows',
 
   /**
    * @virtual
@@ -80,9 +90,30 @@ export default Component.extend(GlobalActions, {
   ]),
 
   /**
+   * @type {String}
+   */
+  activeSlide: undefined,
+
+  /**
    * @type {Boolean}
    */
   isCarouselVisible: true,
+
+  /**
+   * @type {WorkflowEditorViewModificationState}
+   */
+  editorModificationState: undefined,
+
+  /**
+   * If true, then there is "unsaved changes" modal visible
+   * @type {Boolean}
+   */
+  isAskingUserForUnsavedChanges: false,
+
+  /**
+   * @type {Window}
+   */
+  _window: window,
 
   /**
    * @override
@@ -107,14 +138,8 @@ export default Component.extend(GlobalActions, {
   activeAtmWorkflowSchemaIdFromUrl: reads('navigationState.aspectOptions.workflowId'),
 
   /**
-   * @type {ComputedProperty<String>}
+   * @type {ComputedProperty<PromiseObject<Model.AtmWorkflowSchema>>}
    */
-  activeSlide: conditional(
-    array.includes('possibleSlideIds', 'activeSlideFromUrl'),
-    'activeSlideFromUrl',
-    'possibleSlideIds.firstObject'
-  ),
-
   activeAtmWorkflowSchemaProxy: promise.object(computed(
     'activeAtmWorkflowSchemaId',
     'atmInventory',
@@ -145,55 +170,33 @@ export default Component.extend(GlobalActions, {
     }
   )),
 
-  activeAtmWorkflowSchemaIdFromUrlObserver: observer(
-    'activeAtmWorkflowSchemaIdFromUrl',
-    function activeAtmWorkflowSchemaIdFromUrlObserver() {
-      const {
-        activeAtmWorkflowSchemaId,
-        activeAtmWorkflowSchemaIdFromUrl,
-        activeSlide,
-      } = this.getProperties(
-        'activeAtmWorkflowSchemaId',
-        'activeAtmWorkflowSchemaIdFromUrl',
-        'activeSlide'
-      );
+  /**
+   * @type {ComputedProperty<Function>}
+   */
+  routeChangeHandler: computed(function routeChangeHandler() {
+    return (transition) => this.handleRouteChange(transition);
+  }),
 
-      if (activeAtmWorkflowSchemaId !== activeAtmWorkflowSchemaIdFromUrl) {
-        this.set('activeAtmWorkflowSchemaId', activeAtmWorkflowSchemaIdFromUrl);
-        if (['lambdaSelector', 'taskDetails'].includes(activeSlide)) {
-          scheduleOnce('afterRender', this, 'changeSlideViaUrl', 'editor');
-        }
-      }
+  /**
+   * @type {Ember.ComputedProperty<Function>}
+   */
+  pageUnloadHandler: computed(function pageUnloadHandler() {
+    return (event) => this.handlePageUnload(event);
+  }),
+
+  urlParamsObserver: observer(
+    'activeAtmWorkflowSchemaIdFromUrl',
+    'activeSlideFromUrl',
+    function urlParamsObserver() {
+      scheduleOnce('actions', this, 'synchronizeStateWithUrl');
     }
   ),
-
-  activeSlideObserver: observer('activeSlide', function activeSlideObserver() {
-    const scrollableParent = this.$().parents('.ps')[0];
-    if (scrollableParent) {
-      scrollableParent.scroll({
-        top: 0,
-        behavior: 'smooth',
-      });
-    }
-    const activeSlide = this.get('activeSlide');
-    if (!['lambdaSelector', 'lambdaCreator', 'taskDetails'].includes(activeSlide)) {
-      this.cancelTaskDetailsProvider();
-    }
-  }),
 
   atmInventoryObserver: observer('atmInventory', function atmInventoryObserver() {
     // rerender carousel from scratch to avoid animations of slide change
     this.set('isCarouselVisible', false);
     scheduleOnce('afterRender', this, 'set', 'isCarouselVisible', true);
   }),
-
-  changeSlideViaUrl(newSlide, slideParams = {}) {
-    const workflowId = this.get('activeAtmWorkflowSchemaIdFromUrl') || null;
-    this.get('navigationState')
-      .changeRouteAspectOptions(Object.assign({ workflowId }, slideParams, {
-        view: newSlide,
-      }));
-  },
 
   init() {
     this._super(...arguments);
@@ -205,10 +208,255 @@ export default Component.extend(GlobalActions, {
       (...args) => this.runTaskDetailsProvider('edit', ...args)
     );
     this.setProperties({
+      activeSlide: this.getDefaultSlideId(),
       actionsPerSlide: {},
       actionsFactory,
     });
-    this.activeAtmWorkflowSchemaIdFromUrlObserver();
+    this.urlParamsObserver();
+    this.registerRouteChangeHandler();
+    this.registerPageUnloadHandler();
+  },
+
+  willDestroyElement() {
+    try {
+      this.unregisterRouteChangeHandler();
+      this.unregisterPageUnloadHandler();
+    } finally {
+      this._super(...arguments);
+    }
+  },
+
+  registerRouteChangeHandler() {
+    const {
+      routeChangeHandler,
+      router,
+    } = this.getProperties('routeChangeHandler', 'router');
+
+    // Making sure if router has `.on` method (`routeWillChange` hack)
+    // TODO: VFS-8267 Remove this check
+    if (typeof router.on === 'function') {
+      router.on('routeWillChange', routeChangeHandler);
+    }
+  },
+
+  unregisterRouteChangeHandler() {
+    const {
+      routeChangeHandler,
+      router,
+    } = this.getProperties('routeChangeHandler', 'router');
+
+    // Making sure if router has `.on` method (`routeWillChange` hack)
+    // TODO: VFS-8267 Remove this check
+    if (typeof router.off === 'function') {
+      router.off('routeWillChange', routeChangeHandler);
+    }
+  },
+
+  registerPageUnloadHandler() {
+    const {
+      pageUnloadHandler,
+      _window,
+    } = this.getProperties('pageUnloadHandler', '_window');
+
+    _window.addEventListener('beforeunload', pageUnloadHandler);
+  },
+
+  unregisterPageUnloadHandler() {
+    const {
+      pageUnloadHandler,
+      _window,
+    } = this.getProperties('pageUnloadHandler', '_window');
+
+    _window.removeEventListener('beforeunload', pageUnloadHandler);
+  },
+
+  async synchronizeStateWithUrl() {
+    const {
+      activeSlide,
+      activeSlideFromUrl,
+      activeAtmWorkflowSchemaId,
+      activeAtmWorkflowSchemaIdFromUrl,
+    } = this.getProperties(
+      'activeSlide',
+      'activeSlideFromUrl',
+      'activeAtmWorkflowSchemaId',
+      'activeAtmWorkflowSchemaIdFromUrl',
+    );
+
+    let nextActiveSlide = activeSlide;
+    let nextActiveAtmWorkflowSchemaId = activeAtmWorkflowSchemaId || null;
+    const isActiveAtmWorkflowSchemaIdChanged =
+      () => nextActiveAtmWorkflowSchemaId !== activeAtmWorkflowSchemaId;
+    const isActiveSlideChanged = () => nextActiveSlide !== activeSlide;
+
+    // Detect change of `activeAtmWorkflowSchemaId`. `||` condition
+    // is to filter out inequalities like `null !== undefined`.
+    if (activeAtmWorkflowSchemaIdFromUrl !== activeAtmWorkflowSchemaId && (
+        activeAtmWorkflowSchemaIdFromUrl || activeAtmWorkflowSchemaId
+      )) {
+      nextActiveAtmWorkflowSchemaId = activeAtmWorkflowSchemaIdFromUrl || null;
+    }
+
+    // Detect change of `activeSlide`
+    if (!this.isSlideIdValid(activeSlideFromUrl)) {
+      nextActiveSlide = this.getDefaultSlideId();
+    } else if (activeSlideFromUrl !== activeSlide) {
+      nextActiveSlide = activeSlideFromUrl;
+    }
+
+    // If `activeAtmWorkflowSchemaId` is going to change, then new `activeSlide`
+    // value should be suitable for that change e.g. some editor-related slides
+    // should not be available.
+    if (isActiveAtmWorkflowSchemaIdChanged()) {
+      nextActiveSlide =
+        this.getNextSlideIdOnActiveSchemaChange(nextActiveSlide);
+    }
+
+    // Detect if some workflow schema changes are unsaved and take care of them.
+    const editorGetsTurnedOff = this.isSlideIdUsedByEditor(activeSlide) &&
+      !this.isSlideIdUsedByEditor(nextActiveSlide);
+    const willUpdateClearUnsavedChanges = activeAtmWorkflowSchemaId &&
+      (editorGetsTurnedOff || isActiveAtmWorkflowSchemaIdChanged());
+    if (
+      willUpdateClearUnsavedChanges &&
+      this.shouldBlockTransitionDueToUnsavedChanges()
+    ) {
+      const userDecision = await this.askUserAndProcessUnsavedChanges();
+      switch (userDecision) {
+        case 'keepEditing':
+          // User wants to stay in current state.
+          // Reset url params back to represent actual state of the component
+          this.setUrlParams({
+            view: activeSlide,
+            workflowId: activeAtmWorkflowSchemaId,
+          }, true);
+
+          // Nothing more to do - user has chosen to abort the transiton.
+          return;
+        case 'alreadyAsked':
+          // User triggered next url change when the previous one has not been
+          // commited yet. We have to wait for the first one and ignore current
+          // change.
+          return;
+      }
+    }
+
+    // Perform update of component properties
+    const propsToUpdate = {};
+    if (isActiveSlideChanged()) {
+      propsToUpdate.activeSlide = nextActiveSlide;
+    }
+    if (isActiveAtmWorkflowSchemaIdChanged()) {
+      propsToUpdate.activeAtmWorkflowSchemaId = nextActiveAtmWorkflowSchemaId;
+    }
+    if (Object.keys(propsToUpdate).length) {
+      this.setProperties(propsToUpdate);
+    }
+
+    // Introduce some post-update adjustments related to changes.
+    if (isActiveSlideChanged()) {
+      this.scrollTop();
+      if (!this.isSlideIdUsedByTaskDetailsProvider(nextActiveSlide)) {
+        this.cancelTaskDetailsProvider();
+      }
+    }
+
+    // If url params values are different than used by the component,
+    // then url params should be redefined to ensure values consistency.
+    // Using this.get to retrieve the most recent values, as these could change
+    // while asking user what to do with changes.
+    if (
+      nextActiveSlide !== this.get('activeSlideFromUrl') ||
+      nextActiveAtmWorkflowSchemaId !== this.get('activeAtmWorkflowSchemaIdFromUrl')
+    ) {
+      this.setUrlParams({
+        view: nextActiveSlide,
+        workflowId: nextActiveAtmWorkflowSchemaId,
+      }, true);
+    }
+  },
+
+  async handleRouteChange(transition) {
+    if (this.shouldBlockTransitionDueToUnsavedChanges()) {
+      transition.abort();
+      const userDecision = await this.askUserAndProcessUnsavedChanges();
+      if (userDecision === 'save' || userDecision === 'ignore') {
+        transition.retry();
+      }
+    }
+  },
+
+  handlePageUnload(event) {
+    if (this.shouldBlockTransitionDueToUnsavedChanges()) {
+      return preventPageUnload(event, String(this.t('confirmPageClose')));
+    }
+  },
+
+  shouldBlockTransitionDueToUnsavedChanges() {
+    const isActiveSlideUsedByEditor =
+      this.isSlideIdUsedByEditor(this.get('activeSlide'));
+    const isModified = this.get('editorModificationState.isModified');
+    const isActiveAtmWorkflowSchemaProxyFulfilled =
+      this.cacheFor('activeAtmWorkflowSchemaProxy') &&
+      this.get('activeAtmWorkflowSchemaProxy.isFulfilled');
+
+    return isModified &&
+      isActiveSlideUsedByEditor &&
+      isActiveAtmWorkflowSchemaProxyFulfilled;
+  },
+
+  /**
+   * @returns {Promise<String>} one of: `'ignore'`, `'save'`, `'keepEditing'`,
+   * `'alreadyAsked'`
+   */
+  async askUserAndProcessUnsavedChanges() {
+    if (this.get('isAskingUserForUnsavedChanges')) {
+      // User is already in the middle of choosing what to do. It means, that
+      // there was some uncommited url/route change earlier, which needs to
+      // be resolved at the first place.
+      return 'alreadyAsked';
+    }
+    this.set('isAskingUserForUnsavedChanges', true);
+    const executeSaveAction =
+      this.get('editorModificationState.executeSaveAction');
+
+    let decision = 'keepEditing';
+    await this.get('modalManager').show('unsaved-changes-question-modal', {
+      onSubmit: async ({ shouldSaveChanges }) => {
+        if (shouldSaveChanges && executeSaveAction) {
+          const saveResult = await executeSaveAction();
+          if (saveResult && get(saveResult, 'status') === 'failed') {
+            // In case of failure `executeSaveAction` should show proper error
+            // notification. After fail user should stay in editor view and
+            // decide what to do next. Hence `decision` is `'keepEditing'`.
+            return;
+          }
+          decision = 'save';
+        } else {
+          decision = 'ignore';
+        }
+      },
+    }).hiddenPromise;
+    safeExec(this, () => {
+      this.set('isAskingUserForUnsavedChanges', false);
+      if (decision !== 'keepEditing') {
+        this.set('editorModificationState.isModified', false);
+      }
+    });
+
+    return decision;
+  },
+
+  setUrlParams(params, replaceHistory = false) {
+    this.get('navigationState')
+      .changeRouteAspectOptions(params, replaceHistory);
+  },
+
+  changeSlideViaUrl(newSlide, slideParams = {}) {
+    const workflowId = this.get('activeAtmWorkflowSchemaIdFromUrl') || null;
+    this.setUrlParams(Object.assign({ workflowId }, slideParams, {
+      view: newSlide,
+    }));
   },
 
   runTaskDetailsProvider(mode, { stores, task }) {
@@ -255,6 +503,42 @@ export default Component.extend(GlobalActions, {
   cancelTaskDetailsProvider() {
     const onFailureCallback = this.get('taskDetailsProviderData.onFailure');
     onFailureCallback && onFailureCallback();
+  },
+
+  getDefaultSlideId() {
+    return 'list';
+  },
+
+  isSlideIdValid(slideId) {
+    return this.get('possibleSlideIds').includes(slideId);
+  },
+
+  isSlideIdUsedByTaskDetailsProvider(slideId) {
+    return [
+      'lambdaSelector',
+      'lambdaCreator',
+      'taskDetails',
+    ].includes(slideId);
+  },
+
+  isSlideIdUsedByEditor(slideId) {
+    return slideId === 'editor' ||
+      this.isSlideIdUsedByTaskDetailsProvider(slideId);
+  },
+
+  getNextSlideIdOnActiveSchemaChange(activeSlideId) {
+    return this.isSlideIdUsedByTaskDetailsProvider(activeSlideId) ?
+      'editor' : activeSlideId;
+  },
+
+  scrollTop() {
+    const scrollableParent = this.$() && this.$().parents('.ps')[0];
+    if (scrollableParent) {
+      scrollableParent.scroll({
+        top: 0,
+        behavior: 'smooth',
+      });
+    }
   },
 
   actions: {
@@ -344,6 +628,9 @@ export default Component.extend(GlobalActions, {
     taskProviderCancel() {
       this.cancelTaskDetailsProvider();
       this.changeSlideViaUrl('editor');
+    },
+    editorModificationStateChange(editorModificationState) {
+      this.set('editorModificationState', editorModificationState);
     },
   },
 });