+++ src/app/components/embedded-content-space-datasets.js
@@ -10,8 +10,6 @@
 import OneproviderEmbeddedContainer from 'onezone-gui/components/oneprovider-embedded-container';
 import layout from 'onezone-gui/templates/components/one-embedded-container';
 import { inject as service } from '@ember/service';
-import { computed } from '@ember/object';
-import { reads } from '@ember/object/computed';
 import EmbeddedBrowserCommon from 'onezone-gui/mixins/embedded-browser-common';
 import notImplementedWarn from 'onedata-gui-common/utils/not-implemented-warn';
 
@@ -64,9 +62,18 @@ export default OneproviderEmbeddedContainer.extend(EmbeddedBrowserCommon, {
   viewMode: undefined,
 
   /**
-   * @override
+   * **Injected to embedded iframe.**
+   * @virtual
+   * @type {String}
    */
-  embeddedBrowserType: 'datasets',
+  dirId: undefined,
+
+  /**
+   * **Injected to embedded iframe.**
+   * @virtual
+   * @type {Array<String>}
+   */
+  selected: undefined,
 
   /**
    * Dataset state tree to show. One of: attached, detached.
@@ -77,30 +84,9 @@ export default OneproviderEmbeddedContainer.extend(EmbeddedBrowserCommon, {
   attachmentState: undefined,
 
   /**
-   * **Injected to embedded iframe.**
-   * @type {string}
+   * @override
    */
-  dirId: reads('navigationState.aspectOptions.dir'),
-
-  // TODO: VFS-7633 redundancy; create computed util for getting array from aspectOptions
-  /**
-   * List of dataset entity ids that are selected
-   * 
-   * **Injected to embedded iframe.**
-   * @type {Array<String>}
-   */
-  selected: computed('navigationState.aspectOptions.selected.[]', {
-    get() {
-      const rawSelected = this.get('navigationState.aspectOptions.selected');
-      return rawSelected && rawSelected.split(',') || [];
-    },
-    set(key, value) {
-      this.get('navigationState').changeRouteAspectOptions({
-        selected: value && value.join(',') || null,
-      });
-      return value;
-    },
-  }),
+  embeddedBrowserType: 'datasets',
 
   /**
    * @override implements OneEmbeddedContainer